# perf(es/minifier): Use rope-like structure for the name mangler