--- a/crates/swc_ecma_transforms_base/src/rename/analyzer/rope_map.rs
+++ b/crates/swc_ecma_transforms_base/src/rename/analyzer/rope_map.rs
@@ -1,0 +1,14 @@
+use rustc_hash::FxHashMap;
+
+#[derive(Debug, Clone)]
+pub(super) struct RopeMap<K, V> {
+    inner: Vec<FxHashMap<K, V>>,
+}
+
+impl<K, V> Default for RopeMap<K, V> {
+    fn default() -> Self {
+        Self {
+            inner: Default::default(),
+        }
+    }
+}