--- a/crates/swc_ecma_transforms_base/src/rename/analyzer/rope_map.rs
+++ b/crates/swc_ecma_transforms_base/src/rename/analyzer/rope_map.rs
@@ -6,3 +6,10 @@ use super::scope::{FastId, FastJsWord};
 pub(super) struct ReverseMap {
     inner: Vec<FxHashMap<FastJsWord, Vec<FastId>>>,
 }
+
+impl ReverseMap {
+    pub fn push_entry(&mut self, key: FastJsWord, id: FastId) {
+        let map = self.inner.last_mut().unwrap();
+        map.entry(key).or_default().push(id);
+    }
+}