--- a/crates/swc_ecma_transforms_base/src/rename/analyzer/rope_map.rs
+++ b/crates/swc_ecma_transforms_base/src/rename/analyzer/rope_map.rs
@@ -2,22 +2,15 @@ use rustc_hash::FxHashMap;
 
 use super::scope::{FastId, FastJsWord};
 
-#[derive(Debug)]
-pub(super) struct ReverseMap {
-    inner: Vec<FxHashMap<FastJsWord, Vec<FastId>>>,
+#[derive(Debug, Default)]
+pub(super) struct ReverseMap<'a> {
+    prev: Option<&'a ReverseMap<'a>>,
+
+    inner: FxHashMap<FastJsWord, Vec<FastId>>,
 }
 
-impl Default for ReverseMap {
-    fn default() -> Self {
-        Self {
-            inner: vec![Default::default()],
-        }
-    }
-}
-
-impl ReverseMap {
+impl ReverseMap<'_> {
     pub fn push_entry(&mut self, key: FastJsWord, id: FastId) {
-        let map = self.inner.last_mut().unwrap();
-        map.entry(key).or_default().push(id);
+        self.inner.entry(key).or_default().push(id);
     }
 }