--- a/crates/swc_ecma_transforms_base/src/rename/analyzer/scope.rs
+++ b/crates/swc_ecma_transforms_base/src/rename/analyzer/scope.rs
@@ -185,7 +185,7 @@ impl Scope {
                 continue;
             }
 
-            if self.data.all.contains(&left) {
+            if self.data.all.contains(left) {
                 return false;
             }
         }